Restructure project and add example codes
