+++ ProtobufDeserializer.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.645
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProtobufDeserializer", "src\ProtobufDeserializer\ProtobufDeserializer.csproj", "{2124DE2D-507F-4AD5-84CB-44F04463D31B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Examples", "src\Examples\Examples.csproj", "{08A2C4DD-D125-42F6-905B-AD6C8FF0FC86}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2124DE2D-507F-4AD5-84CB-44F04463D31B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2124DE2D-507F-4AD5-84CB-44F04463D31B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2124DE2D-507F-4AD5-84CB-44F04463D31B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2124DE2D-507F-4AD5-84CB-44F04463D31B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{08A2C4DD-D125-42F6-905B-AD6C8FF0FC86}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{08A2C4DD-D125-42F6-905B-AD6C8FF0FC86}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{08A2C4DD-D125-42F6-905B-AD6C8FF0FC86}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{08A2C4DD-D125-42F6-905B-AD6C8FF0FC86}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {01F6AA18-70DC-4F98-9D89-8E6F1D80C9D6}
+	EndGlobalSection
+EndGlobal
