+++ src/Examples/Program.cs
@@ -1,0 +1,53 @@
+﻿using System;
+using System.IO;
+using ProtobufDeserialiser;
+
+namespace Examples
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            // Protobuf data stream for the Customer object
+            var customerDataStream = "8,1,18,5,75,97,119,97,105,26,4,87,111,110,103,32,1".Split(',');
+            // Customer descriptor stream that is generated by protoc
+            var customerDescriptorStream = "0A BE 01 0A 0E 43 75 73 74 6F 6D 65 72 2E 70 72 6F 74 6F 22 A3 01 0A 08 43 75 73 74 6F 6D 65 72 12 0E 0A 02 49 64 18 01 20 01 28 05 52 02 49 64 12 1C 0A 09 46 69 72 73 74 4E 61 6D 65 18 02 20 01 28 09 52 09 46 69 72 73 74 4E 61 6D 65 12 18 0A 07 53 75 72 6E 61 6D 65 18 03 20 01 28 09 52 07 53 75 72 6E 61 6D 65 12 2A 0A 04 54 79 70 65 18 04 20 01 28 0E 32 16 2E 43 75 73 74 6F 6D 65 72 2E 43 75 73 74 6F 6D 65 72 54 79 70 65 52 04 54 79 70 65 22 23 0A 0C 43 75 73 74 6F 6D 65 72 54 79 70 65 12 0A 0A 06 4E 6F 72 6D 61 6C 10 00 12 07 0A 03 56 49 50 10 01 62 06 70 72 6F 74 6F 33".Split(' ');
+
+            var descStream = new MemoryStream();
+            foreach (var s in customerDescriptorStream)
+            {
+                var b = Convert.ToByte(s, 16);
+                descStream.WriteByte(b);
+            }
+
+            var customerStream = new MemoryStream();
+            foreach (var s in customerDataStream)
+            {
+                var b = Convert.ToByte(s);
+                customerStream.WriteByte(b);
+            }
+
+            Console.WriteLine("Hello World!");
+            var instance = Deserializer.Deserialize(descStream.ToArray(), customerStream.ToArray(), typeof(Customer));
+            var instance2 = Deserializer.Deserialize<Customer>(descStream.ToArray(), customerStream.ToArray());
+
+            Console.ReadLine();
+        }
+    }
+
+    public class Customer
+    {
+        public int Id { get; set; }
+        public string FirstName { get; set; }
+        public string Surname { get; set; }
+        // An extra field that should get ignored...
+        public string LastName { get; set; }
+        public CustomerType Type { get; set; }
+    }
+
+    public enum CustomerType
+    {
+        Normal = 0,
+        Vip
+    }
+}
