--- src/protobuf-deserialiser.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <RootNamespace>ProtobufDeserialiser</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Google.Protobuf" Version="3.11.2" />
-  </ItemGroup>
-
-</Project>
